refactor: modular architecture decomposition for large source files
