+++ .antigravitycli/04426d9f-7dc8-437f-bc57-90e20b0fe115.json
@@ -1,0 +1,1 @@
+/Users/raunak/.gemini/config/projects/04426d9f-7dc8-437f-bc57-90e20b0fe115.json
