+++ .antigravitycli/def82744-276f-443c-8b4b-a512b8b82b50.json
@@ -1,0 +1,1 @@
+/Users/raunak/.gemini/config/projects/def82744-276f-443c-8b4b-a512b8b82b50.json
