+++ aura-core/src/orchestrator/event_handlers.rs
@@ -1,0 +1,203 @@
+use super::structs::{Event, Orchestrator};
+use crate::task::DownloadPhase;
+use crate::{Result, TaskId};
+use tracing::{debug, error, info};
+
+impl Orchestrator {
+    pub(crate) async fn handle_bt_metadata_received(
+        &mut self,
+        meta_id: TaskId,
+        sub_id: TaskId,
+        torrent: crate::torrent::Torrent,
+    ) -> Result<()> {
+        if let Some(bt_task) = self.bt_tasks.get(&sub_id) {
+            bt_task.state.mature(torrent.clone()).await;
+
+            let metadata = crate::worker::Metadata {
+                final_uri: format!("magnet:?xt={}", bt_task.state.info_hash.to_magnet_urn()),
+                total_length: Some(torrent.total_length()),
+                name: Some(torrent.info.name.clone()),
+            };
+            self.handle_subtask_matured(meta_id, sub_id, metadata)
+                .await?;
+        }
+        Ok(())
+    }
+
+    pub(crate) async fn handle_subtask_matured(
+        &mut self,
+        meta_id: TaskId,
+        sub_id: TaskId,
+        metadata: crate::worker::Metadata,
+    ) -> Result<()> {
+        if let Some(meta_task) = self.tasks.get_mut(&meta_id) {
+            let mut needs_reregister = false;
+
+            if meta_task.total_length == 0 {
+                if let Some(len) = metadata.total_length {
+                    info!(%meta_id, %len, "Metadata matured: task initialized");
+                    meta_task.total_length = len;
+                    meta_task.generate_ranges(16); // Default 16 segments
+                    needs_reregister = true;
+                }
+            }
+
+            // Update name if currently unnamed
+            if (meta_task.name == "unnamed" || meta_task.name.is_empty()) && metadata.name.is_some()
+            {
+                let new_name = metadata.name.clone().unwrap();
+                info!(%meta_id, %new_name, "Updating task name from metadata");
+                meta_task.name = new_name;
+                needs_reregister = true;
+            }
+
+            if needs_reregister {
+                // Update storage engine
+                let download_dir = {
+                    let config = self.config.load();
+                    config.storage.download_dir.clone()
+                };
+                let path = std::path::Path::new(&download_dir).join(&meta_task.name);
+                let _ = self
+                    .storage_tx
+                    .send(crate::storage::StorageRequest::RegisterTask {
+                        task_id: meta_id,
+                        path,
+                        total_length: meta_task.total_length,
+                        checksum: meta_task.checksum.clone(),
+                    })
+                    .await;
+            }
+
+            if let Some(sub_task) = meta_task.subtasks.iter_mut().find(|s| s.id == sub_id) {
+                sub_task.phase = DownloadPhase::Downloading;
+            }
+        }
+
+        let (should_notify, final_uri, total_length, name) =
+            if let Some(meta_task) = self.tasks.get(&meta_id) {
+                (
+                    meta_task.total_length > 0,
+                    metadata.final_uri,
+                    meta_task.total_length,
+                    metadata.name,
+                )
+            } else {
+                (false, String::new(), 0, None)
+            };
+
+        if should_notify {
+            let _ = self.event_tx.send(Event::MetadataResolved {
+                id: meta_id,
+                final_uri,
+                total_length,
+                name,
+            });
+        }
+
+        self.dispatch_next_ranges(meta_id, sub_id).await
+    }
+
+    pub(crate) async fn handle_range_finished(
+        &mut self,
+        meta_id: TaskId,
+        sub_id: TaskId,
+        range: crate::task::Range,
+    ) -> Result<()> {
+        let mut completed = false;
+        if let Some(meta_task) = self.tasks.get_mut(&meta_id) {
+            // Racing coordination: check if other subtasks were also working on this range
+            let racing_sub_ids: Vec<TaskId> = meta_task
+                .in_flight_ranges
+                .iter()
+                .filter(|(sid, r)| *r == range && *sid != sub_id)
+                .map(|(sid, _)| *sid)
+                .collect();
+
+            if !racing_sub_ids.is_empty() {
+                debug!(%meta_id, ?range, racing = racing_sub_ids.len(), "Range finished; canceling racing workers");
+                for racing_sid in racing_sub_ids {
+                    // For non-BT tasks, we rely on the in_flight_ranges cleanup and next loop check.
+                    if let Some(sub) = meta_task.subtasks.iter_mut().find(|s| s.id == racing_sid) {
+                        sub.assigned_ranges.retain(|r| *r != range);
+                    }
+                }
+            }
+
+            meta_task.mark_range_complete(sub_id, range);
+
+            if meta_task.is_complete()
+                && meta_task.phase != DownloadPhase::Verifying
+                && meta_task.phase != DownloadPhase::Complete
+            {
+                if meta_task.checksum.is_some() {
+                    info!(%meta_id, "All ranges complete for MetaTask, entering Verifying phase");
+                    meta_task.phase = DownloadPhase::Verifying;
+                } else {
+                    info!(%meta_id, "All ranges complete for MetaTask, entering seeding phase");
+                    meta_task.phase = DownloadPhase::Complete;
+                    if meta_task.seeding_start_time.is_none() {
+                        meta_task.seeding_start_time = Some(chrono::Utc::now());
+                    }
+                }
+                completed = true;
+            }
+        }
+
+        if completed {
+            let _ = self
+                .storage_tx
+                .send(crate::storage::StorageRequest::Complete(meta_id))
+                .await;
+        }
+
+        self.dispatch_next_ranges(meta_id, sub_id).await
+    }
+
+    pub(crate) async fn handle_storage_event(
+        &mut self,
+        event: crate::storage::StorageEvent,
+    ) -> Result<()> {
+        match event {
+            crate::storage::StorageEvent::Completed(id) => {
+                info!(%id, "Storage reported completion");
+                if let Some(task) = self.tasks.get_mut(&id) {
+                    task.phase = DownloadPhase::Complete;
+                    if task.seeding_start_time.is_none() {
+                        task.seeding_start_time = Some(chrono::Utc::now());
+                    }
+                    let _ = self.event_tx.send(Event::TaskProgress {
+                        id,
+                        completed_bytes: task.total_length,
+                        uploaded_bytes: task.uploaded_length,
+                        total_bytes: task.total_length,
+                    });
+                }
+                let event = Event::TaskCompleted(id);
+                let _ = self.event_tx.send(event.clone());
+                self.hook_manager.handle_event(&event).await;
+            }
+            crate::storage::StorageEvent::Error(id, err) => {
+                error!(%id, %err, "Storage reported fatal error; pausing task");
+                let mut exists = false;
+                if let Some(task) = self.tasks.get_mut(&id) {
+                    task.phase = DownloadPhase::Error;
+                    exists = true;
+                }
+
+                if exists {
+                    // Trigger pause logic to cleanup workers
+                    let _ = self.handle_pause(id).await;
+
+                    let event = Event::TaskError {
+                        id,
+                        message: format!("Storage Error: {}", err),
+                    };
+                    let _ = self.event_tx.send(event.clone());
+                    self.hook_manager.handle_event(&event).await;
+                }
+            }
+        }
+        Ok(())
+    }
+}
