+++ aura-core/src/orchestrator/mod.rs
@@ -1,7 +1,11 @@
 pub mod commands;
 pub mod engine;
+pub mod event_handlers;
 pub mod events;
 pub mod lifecycle;
-pub mod logic;
+pub mod monitors;
+pub mod runner;
+pub mod structs;
+
 pub use engine::Engine;
-pub use logic::*;
+pub use structs::{Command, Event, Orchestrator, SubTaskEvent, WorkerCommand};
