+++ aura-core/src/orchestrator/structs.rs
@@ -4,7 +4,6 @@ use crate::bt_worker::PeerId;
 use crate::buffer_pool::BufferPool;
 use crate::dht::DhtCommand;
 use crate::nat::NatCommand;
-use crate::storage::StorageRequest;
 use crate::task::{MetaTask, Range};
 use crate::throttler::Throttler;
 use crate::worker::Metadata;
@@ -12,13 +11,9 @@ use crate::{InfoHash, Result, TaskId};
 use arc_swap::ArcSwap;
 use rand::RngExt;
 use std::collections::HashMap;
-use std::net::{IpAddr, Ipv4Addr};
 use std::sync::Arc;
 use tokio::sync::{broadcast, mpsc};
 use tokio_util::sync::CancellationToken;
-
-pub use super::engine::Engine;
-use super::lifecycle;
 
 /// Internal commands for the Orchestrator.
 #[derive(Debug)]
@@ -99,12 +94,11 @@ pub struct Orchestrator {
     pub(crate) tasks: HashMap<TaskId, MetaTask>,
     pub(crate) bt_registry: HashMap<InfoHash, TaskId>,
     pub(crate) bt_tasks: HashMap<TaskId, Arc<BtTask>>,
-    // Key: sub_id
-    pub(crate) worker_command_txs: HashMap<TaskId, tokio::sync::broadcast::Sender<WorkerCommand>>, // Key: sub_id
+    pub(crate) worker_command_txs: HashMap<TaskId, tokio::sync::broadcast::Sender<WorkerCommand>>,
     pub(crate) cancellation_tokens: HashMap<TaskId, CancellationToken>,
     pub(crate) command_rx: mpsc::Receiver<Command>,
     pub(crate) event_tx: broadcast::Sender<Event>,
-    pub(crate) storage_tx: mpsc::Sender<StorageRequest>,
+    pub(crate) storage_tx: mpsc::Sender<crate::storage::StorageRequest>,
     pub(crate) storage_completion_rx: mpsc::Receiver<crate::storage::StorageEvent>,
     pub(crate) subtask_tx: mpsc::Sender<SubTaskEvent>,
     pub(crate) subtask_rx: mpsc::Receiver<SubTaskEvent>,
@@ -221,9 +215,6 @@ impl Orchestrator {
                 crate::vpn::VpnStatus::Connected => Ok(()),
                 crate::vpn::VpnStatus::Connecting => {
                     tracing::info!("VPN is connecting, waiting...");
-                    // Simple retry loop or just wait?
-                    // For now, let's just fail and let the next attempt try again
-                    // or let auto_connect handle it.
                     Err(crate::Error::Engine("VPN is still connecting".to_string()))
                 }
                 crate::vpn::VpnStatus::Disconnected | crate::vpn::VpnStatus::Error(_) => {
@@ -242,9 +233,8 @@ impl Orchestrator {
                 }
             }
         } else {
-            // Force tunnel is on but no provider is configured
             if config.network.interface.is_some() {
-                Ok(()) // Interface monitor is technically a provider, handled above
+                Ok(())
             } else {
                 Err(crate::Error::Config(
                     "Mandatory tunnel enabled but no VPN provider or interface configured"
@@ -257,7 +247,7 @@ impl Orchestrator {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         command_rx: mpsc::Receiver<Command>,
-        storage_tx: mpsc::Sender<StorageRequest>,
+        storage_tx: mpsc::Sender<crate::storage::StorageRequest>,
         storage_completion_rx: mpsc::Receiver<crate::storage::StorageEvent>,
         dht_tx: mpsc::Sender<DhtCommand>,
         lpd_tx: mpsc::Sender<crate::lpd::LpdCommand>,
@@ -328,205 +318,5 @@ impl Orchestrator {
             },
             event_tx,
         )
-    }
-
-    pub(crate) async fn perform_adaptive_scaling(&mut self) {
-        let config = self.config.load();
-        let max_concurrency = config.bandwidth.max_connections_per_task;
-
-        // EWMA factor
-        let alpha = 0.3;
-
-        let mut to_dispatch = Vec::new();
-
-        for task in self.tasks.values_mut() {
-            if task.phase != crate::task::DownloadPhase::Downloading {
-                continue;
-            }
-
-            for sub_task in task.subtasks.iter_mut() {
-                if !sub_task.active {
-                    continue;
-                }
-
-                // Calculate throughput for the last second
-                let current_throughput = sub_task.recent_bytes_downloaded as f64;
-                sub_task.recent_bytes_downloaded = 0;
-
-                // Update EWMA
-                if sub_task.ewma_throughput == 0.0 {
-                    sub_task.ewma_throughput = current_throughput;
-                } else {
-                    sub_task.ewma_throughput =
-                        (alpha * current_throughput) + ((1.0 - alpha) * sub_task.ewma_throughput);
-                }
-
-                // Adaptive Scaling Logic
-                // If throughput per connection is low (< 256 KB/s) and we haven't reached max_concurrency, scale up.
-                let throughput_per_connection = if sub_task.target_concurrency > 0 {
-                    sub_task.ewma_throughput / sub_task.target_concurrency as f64
-                } else {
-                    0.0
-                };
-
-                if sub_task.assigned_ranges.len() < sub_task.target_concurrency {
-                    to_dispatch.push((task.id, sub_task.id));
-                } else if throughput_per_connection < 256.0 * 1024.0
-                    && sub_task.target_concurrency < max_concurrency
-                {
-                    sub_task.target_concurrency =
-                        (sub_task.target_concurrency + 1).min(max_concurrency);
-                    tracing::debug!(
-                        meta_id = %task.id,
-                        sub_id = %sub_task.id,
-                        target = %sub_task.target_concurrency,
-                        throughput = %sub_task.ewma_throughput,
-                        "Scaling up subtask concurrency due to low throughput"
-                    );
-
-                    to_dispatch.push((task.id, sub_task.id));
-                }
-            }
-        }
-
-        for (meta_id, sub_id) in to_dispatch {
-            // Re-dispatch to spawn new workers immediately
-            let _ = self.dispatch_next_ranges(meta_id, sub_id).await;
-        }
-    }
-
-    pub async fn run(mut self) -> Result<()> {
-        tracing::info!("Orchestrator started");
-
-        let local_addr = self.resolve_local_addr();
-        let config_initial = self.config.load();
-        let bind_addr = std::net::SocketAddr::new(
-            local_addr.unwrap_or(IpAddr::V4(Ipv4Addr::UNSPECIFIED)),
-            config_initial.network.listen_port,
-        );
-
-        let listener = tokio::net::TcpListener::bind(bind_addr)
-            .await
-            .map_err(|e| {
-                crate::Error::Config(format!(
-                    "Failed to bind Peer Listener on {}: {}",
-                    bind_addr, e
-                ))
-            })?;
-        tracing::info!("Peer Listener listening on {}", bind_addr);
-
-        let mut save_interval = tokio::time::interval(std::time::Duration::from_secs(
-            config_initial.storage.save_session_interval_secs,
-        ));
-        let mut scaling_interval = tokio::time::interval(std::time::Duration::from_secs(1));
-
-        // VPN Kill-switch Monitor
-        let vpn_watch_rx = self.vpn_watch_tx.subscribe();
-        let subtask_tx_monitor = self.subtask_tx.clone();
-        let config_monitor = self.config.clone();
-
-        tokio::spawn(async move {
-            let mut interval = tokio::time::interval(std::time::Duration::from_secs(2));
-            loop {
-                interval.tick().await;
-
-                let force_tunnel = config_monitor.load().vpn.force_tunnel;
-
-                // Pick up the latest provider
-                let vpn_opt = vpn_watch_rx.borrow().clone();
-
-                if let Some(vpn) = vpn_opt {
-                    if force_tunnel {
-                        match vpn.status().await {
-                            Ok(crate::vpn::VpnStatus::Disconnected)
-                            | Ok(crate::vpn::VpnStatus::Error(_)) => {
-                                tracing::warn!(
-                                    provider = %vpn.name(),
-                                    interface = ?vpn.interface(),
-                                    "VPN Kill-switch triggered! Connection lost. Stopping all tasks."
-                                );
-                                if subtask_tx_monitor
-                                    .send(SubTaskEvent::KillSwitch)
-                                    .await
-                                    .is_err()
-                                {
-                                    break;
-                                }
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-
-                if subtask_tx_monitor.is_closed() {
-                    break;
-                }
-            }
-        });
-
-        loop {
-            tokio::select! {
-                _ = scaling_interval.tick() => {
-                    self.perform_adaptive_scaling().await;
-                }
-                _ = save_interval.tick() => {
-                    self.check_seed_limits().await;
-                    let ids: Vec<TaskId> = self.tasks.keys().cloned().collect();
-                    for id in ids {
-                        let _ = self.save_task(id).await;
-                    }
-                }
-                Ok((stream, addr)) = listener.accept() => {
-                    let bt_registry = self.bt_registry.clone();
-                    let worker_command_txs = self.worker_command_txs.clone();
-                    let storage_tx = self.storage_tx.clone();
-                    let subtask_tx = self.subtask_tx.clone();
-                    let my_peer_id = self.peer_id;
-                    let cancellation_tokens = self.cancellation_tokens.clone();
-                    let local_addr = self.resolve_local_addr();
-                    let config = self.config.load().clone();
-                    let pool = self.pool.clone();
-                    let throttler = self.throttler.clone();
-
-                    let bt_tasks = self.bt_tasks.clone();
-
-                    tokio::spawn(async move {
-                        if let Err(e) = lifecycle::handle_incoming_peer(stream, addr, bt_registry, bt_tasks, worker_command_txs, storage_tx, subtask_tx, my_peer_id, cancellation_tokens, local_addr, config, pool, throttler).await {
-                            tracing::debug!(?addr, error = %e, "Failed to handle incoming peer");
-                        }
-                    });
-                }
-                Some(event) = self.subtask_rx.recv() => {
-                    if let Err(e) = self.handle_subtask_event(event).await {
-                        tracing::error!("Event handle error: {}", e);
-                    }
-                    self.update_power_management();
-                }
-                cmd_res = self.command_rx.recv() => {
-                    match cmd_res {
-                        Some(cmd) => {
-                            if let Err(e) = self.handle_command(cmd).await {
-                                if e.to_string().contains("Shutting down") {
-                                    tracing::info!("Orchestrator shutting down gracefully");
-                                    return Ok(());
-                                }
-                                tracing::error!("Command handle error: {}", e);
-                            }
-                        }
-                        None => {
-                            tracing::warn!("Orchestrator command channel closed, exiting loop");
-                            return Ok(());
-                        }
-                    }
-                    self.update_power_management();
-                }
-                Some(event) = self.storage_completion_rx.recv() => {
-                    if let Err(e) = self.handle_storage_event(event).await {
-                        tracing::error!("Storage event error: {}", e);
-                    }
-                    self.update_power_management();
-                }
-            }
-        }
     }
 }
