+++ aura-daemon/src/assets.rs
@@ -1,0 +1,32 @@
+use axum::{
+    body::Body,
+    extract::Path,
+    http::{header, StatusCode},
+    response::{IntoResponse, Response},
+};
+use rust_embed::RustEmbed;
+
+#[derive(RustEmbed)]
+#[folder = "web/"]
+pub struct Assets;
+
+pub async fn index_handler() -> impl IntoResponse {
+    static_handler(Path("index.html".to_string())).await
+}
+
+pub async fn static_handler(Path(path): Path<String>) -> impl IntoResponse {
+    let path = path.trim_start_matches('/');
+    let mime_type = mime_guess::from_path(path).first_or_octet_stream();
+
+    match Assets::get(path) {
+        Some(content) => Response::builder()
+            .status(StatusCode::OK)
+            .header(header::CONTENT_TYPE, mime_type.as_ref())
+            .body(Body::from(content.data))
+            .unwrap(),
+        None => Response::builder()
+            .status(StatusCode::NOT_FOUND)
+            .body(Body::from("404 Not Found"))
+            .unwrap(),
+    }
+}
